[refactor] Consolidate esmapping-generator code
