+++ internal/storage/v1/elasticsearch/mappings/flags_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package app
+package mappings
 
 import (
 	"testing"
@@ -9,8 +9,6 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
-	"github.com/jaegertracing/jaeger/pkg/testutils"
 )
 
 func TestOptionsWithDefaultFlags(t *testing.T) {
@@ -50,8 +48,4 @@ func TestOptionsWithFlags(t *testing.T) {
 	assert.Equal(t, "test", o.IndexPrefix)
 	assert.Equal(t, "true", o.UseILM)
 	assert.Equal(t, "jaeger-test-policy", o.ILMPolicyName)
-}
-
-func TestMain(m *testing.M) {
-	testutils.VerifyGoLeaks(m)
 }
